+++ codemp/client/cl_ui.cpp
@@ -691,7 +691,7 @@ intptr_t CL_UISystemCalls( intptr_t *args ) {
 		Com_Memcpy( VMA(1), VMA(2), args[3] );
 		return 0;
 	case TRAP_STRNCPY:
-		return (int)strncpy( (char *)VMA(1), (const char *)VMA(2), args[3] );
+		return (intptr_t)strncpy( (char *)VMA(1), (const char *)VMA(2), args[3] );
 	case TRAP_SIN:
 		return FloatAsInt( sin( VMF(1) ) );
 	case TRAP_COS:
@@ -1077,7 +1077,7 @@ intptr_t CL_UISystemCalls( intptr_t *args ) {
 		char *languageName,*holdName;
 
 		holdName = ((char *)VMA(2));
-		languageName = (char *) SE_GetLanguageName((const int)VMA(1));
+		languageName = (char *) SE_GetLanguageName((const intptr_t)VMA(1));
 		Q_strncpyz( holdName, languageName,128 );
 		return 0;
